[SPARK-9057] [STREAMING]
Based on feedback from tdas, the relevant example class names now include the word "join" in order to be clearer that the example shows how to join a streaming RDD; the Java example is reformatted to use 2 spaces for indentation; and, more comments are included.
